test/smoke: move methods to appropriate files
